MNT: Rename example files with 'test' in name
In PR 23130 it was noticed that certain example files had 'test' in their
name and were getting picked up automatically by pytest accidentally. To
avoid this, `testpaths = lib` was added to pytest.ini. While this is sufficient
to avoid future problems, removing the word 'test' from example file names and
replacing it with 'demo', which is commonly used in other matplotlib examples,
provides future safeguards.

To avoid causing broken links in the example documentation
https://matplotlib.org/stable/gallery/index.html
add `redirect-from` statements to the old file names as directed in
https://matplotlib.org/stable/devel/documenting_mpl.html#moving-documentation.